requirement.txt
ffmpeg-python
openai-whisper
transformers
torch
pysrt
